add download_artifacts.py
