+++ scripts/download_artifacts.py
@@ -1,4 +1,8 @@
-"""Download the latest performance artificat from the GitHub actions runs."""
+"""
+    Download the latest performance artifacts from the GitHub actions runs.
+    This is for use to get performance statistics from a GetHub action run and append to an archive file.
+    Call this sometime after the GitHub action has run to archive the data.
+"""
 
 # pylint: disable=C0301,W1514
 
@@ -7,10 +11,11 @@ import os
 import zipfile
 import io
 import requests
+import append_artifacts
 
 
 def main():
-    """Main routine to download artificats and append to archive file csv file."""
+    """Main routine to download artifacts and append to archive file csv file."""
     if len(sys.argv) <= 1:
         print("Specify archive file as command line argument")
         return
@@ -21,7 +26,7 @@ def main():
         print("No performance runs found")
         return
     csv_contents = get_artifact(headers, run_id)
-    append_csv_file(csv_contents, archive_file)
+    append_artifacts.append_csv_file(csv_contents, archive_file)
 
 
 def get_url_headers():
@@ -74,35 +79,6 @@ def get_artifact(headers, run_id):
                 with z.open(file_name) as f:
                     csv_contents = f.read().decode("utf-8")
     return csv_contents
-
-
-def append_csv_file(csv_contents, archive_csv_file):
-    """Append the rows from csv_contents to the archive file for rows that are not already in the file."""
-
-    date_map = {}
-    if os.path.exists(archive_csv_file):
-        # Archive file exists, read file to collect all dates in file
-        with open(archive_csv_file, "r") as fp:
-            file_contents = fp.read()
-            for line in file_contents.split("\n"):
-                columns = line.split(",")
-                date_map[columns[0]] = True
-    else:
-        # Archive file does not exists, create it with a CSV file header
-        with open(archive_csv_file, "a+") as fp:
-            fp.write(
-                "date,scenario,hf_hydrodata_version,hydrodata_url,subsettools_version,remotelocal,server,cpus,hotcold,wy,comment,duration\n"
-            )
-
-    added_rows = 0
-    with open(archive_csv_file, "a+") as fp:
-        for line in csv_contents.split("\n"):
-            if line:
-                columns = line.split(",")
-                if not date_map.get(columns[0]):
-                    fp.write(f"{line}\n")
-                    added_rows = added_rows + 1
-    print(f"Added {added_rows} rows to {archive_csv_file} file.")
 
 
 if __name__ == "__main__":
